2095: AFP - Added Timestamp in the AFP engine
git-svn-id: svn+ssh://svn.amanzi.org/var/lib/svn/trunk/awe@4927 0edadee9-8154-4c59-a661-b5f45207fcbd

+++ org.amanzi.awe.afp/src/org/amanzi/awe/afp/executors/AfpProcessProgress.java
@@ -6,12 +6,12 @@ public interface AfpProcessProgress {
 
 	public void onProgressUpdate(int result, 
 			long time, // time in millisecs
-			int remaingtotal, 
-			int sectorSeperations,
-			int siteSeperation,
-			int freqConstraints,
-			int interference,
-			int neighbor,
-			int tringulation,
-			int shadowing);
+			long remaingtotal, 
+			long sectorSeperations,
+			long siteSeperation,
+			long freqConstraints,
+			long interference,
+			long neighbor,
+			long tringulation,
+			long shadowing);
 }
